--- src/fairseq2/datasets/instruction/base.py
@@ -1,67 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
-#
-# This source code is licensed under the BSD-style license found in the
-# LICENSE file in the root directory of this source tree.
-
-from abc import ABC, abstractmethod
-from typing import Any, Set
-
-from fairseq2.data.text import TextTokenizer
-from fairseq2.datasets.data_reader import DataPipelineReader
-from fairseq2.datasets.loader import DelegatingDatasetLoader
-from fairseq2.gang import Gang
-from fairseq2.models.sequence import SequenceBatch
-
-
-class InstructionDataset(ABC):
-    """Represents an instruction finetuning dataset."""
-
-    @abstractmethod
-    def create_reader(
-        self,
-        split: str,
-        tokenizer: TextTokenizer,
-        gang: Gang,
-        max_seq_len: int,
-        max_num_tokens: int,
-        *,
-        shuffle_window_size: int = 1,
-        num_accumulate: int = 1,
-        num_prefetch: int = 1,
-        seed: int = 2,
-        **extras: Any,
-    ) -> DataPipelineReader[SequenceBatch]:
-        """Create a dataset reader.
-
-        :param split:
-            The split to read.
-        :param tokenizer:
-            The tokenizer to encode text.
-        :param gang:
-            The gang over which to shard the dataset.
-        :param max_seq_len:
-            The maximum sequence length of each example. Examples longer than
-            this value will be dropped.
-        :param max_num_tokens:
-            The maximum number of tokens in each batch.
-        :param shuffle_window_size:
-            The size of the shuffle window. If ``1``, no shuffling is performed;
-            if ``0``, performs true shuffling by loading the entire dataset.
-        :param num_accumulate:
-            The number of batches to accumulate in each iteration. Typically
-            used with gradient accumulation during training.
-        :param num_prefetch:
-            The number of batches to prefetch in background.
-        :param seed:
-            The seed to initialize the random number generators used internally.
-        :param extras:
-            The extra parameters specific to the dataset implementation.
-        """
-
-    @abstractmethod
-    def splits(self) -> Set[str]:
-        """Return the set of splits."""
-
-
-load_instruction_dataset = DelegatingDatasetLoader[InstructionDataset]()
