+++ src/fairseq2/datasets/parallel_text.py
@@ -6,9 +6,21 @@
 
 from __future__ import annotations
 
+from abc import ABC, abstractmethod
 from collections import defaultdict
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Sequence, Set, Tuple, cast, final
+from typing import (
+    Any,
+    Dict,
+    List,
+    NamedTuple,
+    Optional,
+    Sequence,
+    Set,
+    Tuple,
+    cast,
+    final,
+)
 
 from fairseq2.assets import AssetCard, AssetCardError
 from fairseq2.data import (
@@ -19,18 +31,88 @@ from fairseq2.data import (
     create_bucket_sizes,
 )
 from fairseq2.data.text import TextTokenizer, read_text
-from fairseq2.datasets.data_reader import DataPipelineReader
+from fairseq2.datasets.data_reader import DataPipelineReader, DataReader
 from fairseq2.datasets.error import DatasetError
-from fairseq2.datasets.loader import AbstractDatasetLoader
-from fairseq2.datasets.parallel_text.base import (
-    LangPair,
-    ParallelTextDataset,
-    load_parallel_text_dataset,
-)
+from fairseq2.datasets.loader import AbstractDatasetLoader, DelegatingDatasetLoader
 from fairseq2.gang import Gang
 from fairseq2.models.seq2seq import Seq2SeqBatch
 from fairseq2.nn.padding import get_seqs_and_padding_mask
 from fairseq2.typing import Device, override
+
+
+class LangPair(NamedTuple):
+    """Represents the language pair of a parallel corpus."""
+
+    source_lang: str
+    """The source language code."""
+
+    target_lang: str
+    """The target language code."""
+
+    def __repr__(self) -> str:
+        return f"{self.source_lang}-{self.target_lang}"
+
+
+class ParallelTextDataset(ABC):
+    """Represents a parallel text dataset."""
+
+    @abstractmethod
+    def create_reader(
+        self,
+        split: str,
+        tokenizer: TextTokenizer,
+        gang: Gang,
+        max_seq_len: int,
+        max_num_tokens: int,
+        *,
+        lang_pairs: Optional[Sequence[LangPair]] = None,
+        sample: bool = False,
+        shuffle_window_size: int = 1,
+        num_accumulate: int = 1,
+        num_prefetch: int = 1,
+        seed: int = 2,
+        **extras: Any,
+    ) -> DataReader[Seq2SeqBatch]:
+        """Create a dataset reader.
+
+        :param split:
+            The split to read.
+        :param tokenizer:
+            The tokenizer to encode text.
+        :param gang:
+            The gang over which to shard the dataset.
+        :param max_seq_len:
+            The maximum sequence length of each example. Examples longer than
+            this value will be dropped.
+        :param max_num_tokens:
+            The maximum number of tokens in each batch.
+        :param lang_pairs:
+            The language pairs to read. If ``None``, all pairs will be read.
+        :param sample:
+            If ``True``, language pair corpora will be sampled in proportion to
+            their size.
+        :param shuffle_window_size:
+            The size of the shuffle window. If ``1``, no shuffling is performed;
+            if ``0``, performs true shuffling by loading the entire dataset.
+        :param num_accumulate:
+            The number of batches to accumulate in each iteration. Typically
+            used with gradient accumulation during training.
+        :param num_prefetch:
+            The number of batches to prefetch in background.
+        :param seed:
+            The seed to initialize the random number generators used internally.
+        :param extras:
+            The extra parameters specific to the dataset implementation.
+        """
+
+    @abstractmethod
+    def splits(self) -> Set[str]:
+        """Return the set of splits."""
+
+    @abstractmethod
+    def lang_pairs(self, split: str) -> Set[LangPair]:
+        """Return the set of language pairs of ``split``."""
+
 
 # TODO: FIX, INFER
 num_parallel_calls = 10
@@ -360,6 +442,9 @@ class _NllbDataPipelineBuilder:
         )
 
 
+load_parallel_text_dataset = DelegatingDatasetLoader[ParallelTextDataset]()
+
+
 @final
 class NllbDatasetLoader(AbstractDatasetLoader[NllbDataset]):
     """Loads NLLB datasets."""
@@ -393,5 +478,5 @@ class NllbDatasetLoader(AbstractDatasetLoader[NllbDataset]):
 load_nllb_dataset = NllbDatasetLoader()
 
 
-def _register_nllb() -> None:
+def _register_parallel_text() -> None:
     load_parallel_text_dataset.register("nllb", load_nllb_dataset)
